requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Scientific Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Reinforcement Learning
gymnasium>=0.29.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.10.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
scikit-learn>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
